+++ src/arithmetic/ibex_AffineVar.cpp
@@ -1,5 +1,5 @@
 /* ============================================================================
- * I B E X - Implementation of the AffineMain<AF_fAF2> class based on fAF version 2
+ * I B E X - Implementation of the AffineVarMain class
  * ============================================================================
  * Copyright   : ENSTA Bretagne (FRANCE)
  * License     : This program can be distributed under the terms of the GNU LGPL.
@@ -154,7 +154,7 @@ AffineVarMain<AF_fAFFullI>& AffineVarMain<AF_fAFFullI>::operator=(const Interval
 			_actif = 0;
 		} else {
 			_actif = 1;
-			std::pair<int,double> p(_var, x.rad());
+			std::pair<af3_int,double> p(_var, x.rad());
 			_elt._rays.push_back(p);
 		}
 		if (AF_fAFFullI::_counter <= _var ) {AF_fAFFullI::_counter = _var+1;}
@@ -174,9 +174,9 @@ AffineVarMain<AF_fAFFullI>& AffineVarMain<AF_fAFFullI>::operator=(const AffineVa
 		_elt._rays.clear();
 
 		if ((x.is_actif()) && (!x._elt._rays.empty()))	{
-			std::list<std::pair<int,double> >::const_iterator it = x._elt._rays.begin();
+			std::list<std::pair<af3_int,double> >::const_iterator it = x._elt._rays.begin();
 			for (; it != x._elt._rays.end(); ++it) {
-				_elt._rays.push_back(std::pair<int,double>(it->first,it->second));
+				_elt._rays.push_back(std::pair<af3_int,double>(it->first,it->second));
 			}
 		}
 	}
