+++ src/function/ibex_AffineEval.h
@@ -1,12 +1,14 @@
 /* ============================================================================
  * I B E X - Evaluation of Affine forms
  * ============================================================================
+ * Copyright   : ENSTA Bretagne (FRANCE)
  * License     : This program can be distributed under the terms of the GNU LGPL.
- *               See the file COPYING.LESSER.
+ *               See the file LICENCE.
  *
  * Author(s)   : Jordan Ninin
- * Created     : April 08, 2013
+ * Created     : June 6, 2020
  * ---------------------------------------------------------------------------- */
+
 
 #ifndef __IBEX_AFFINE_EVAL_H__
 #define __IBEX_AFFINE_EVAL_H__
